added dtype to torch :)

+++ metta/agent/lib/nn_layer_library.py
@@ -58,7 +58,7 @@ class Swish(nn.Module):
 
     def __init__(self):
         super().__init__()
-        self.beta = nn.Parameter(torch.tensor(1.0))
+        self.beta = nn.Parameter(torch.tensor(1.0), dtype=torch.float32)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         return x * torch.sigmoid(self.beta * x)
